spark/requirements.txt
pyspark==3.5.0
pandas
google-cloud-bigquery